Move logging & messages to private package
