+++ implementation/src/main/java/io/smallrye/config/ImplicitConverters.java
@@ -27,6 +27,7 @@ import java.util.Map;
 
 import org.eclipse.microprofile.config.spi.Converter;
 
+import io.smallrye.config._private.ConfigMessages;
 import io.smallrye.config.common.utils.StringUtil;
 
 /**
