+++ implementation/src/main/java/io/smallrye/config/SecuritySupport.java
@@ -24,6 +24,8 @@ import java.security.PrivilegedActionException;
 import java.security.PrivilegedExceptionAction;
 import java.util.Arrays;
 
+import io.smallrye.config._private.ConfigLogging;
+
 /**
  * @author <a href="http://jmesnil.net/">Jeff Mesnil</a> (c) 2018 Red Hat inc.
  */
