+++ implementation/src/main/java/io/smallrye/config/SmallRyeConfig.java
@@ -45,6 +45,8 @@ import org.eclipse.microprofile.config.spi.Converter;
 
 import io.smallrye.common.annotation.Experimental;
 import io.smallrye.config.SmallRyeConfigBuilder.InterceptorWithPriority;
+import io.smallrye.config._private.ConfigLogging;
+import io.smallrye.config._private.ConfigMessages;
 
 /**
  * @author <a href="http://jmesnil.net/">Jeff Mesnil</a> (c) 2017 Red Hat inc.
